Base version on 'git describe --tags'
This makes the version based on `git describe --tags`, which is
basically `<most recent tag>-<number of commits since that tag>-<commit
hash>`

This is useful for us for development, because it lets us push distinct
images tags and avoid confusion when deploying new manifests on a test
environment.

However, it will take more storage on quay.io.